# release-25.2: gossip: parallelize callback handling